+++ .github/scripts/bazel_build.sh
@@ -63,7 +63,7 @@ BAZEL_BINARY_TARGETS=(
   "//:attention_broker_service"
   "//:attention_broker_client"
   "//:das"
-  "//:tests_db_loader"
+  "//:db_loader"
   "//:busnode"
   "//:busclient"
   "//:database_adapter"
@@ -74,7 +74,7 @@ BAZEL_BINARY_OUTPUTS=(
   "bazel-bin/word_query_evolution"
   "bazel-bin/attention_broker_service"
   "bazel-bin/attention_broker_client"
-  "bazel-bin/tests_db_loader"
+  "bazel-bin/db_loader"
   "bazel-bin/busnode"
   "bazel-bin/busclient"
   "bazel-bin/database_adapter"
