+++ src/package/BUILD
@@ -13,7 +13,7 @@ pkg_files(
         "//:busclient",
         "//:busnode",
         "//:database_adapter",
-        "//:tests_db_loader",
+        "//:db_loader",
         "//:word_query",
         "//:word_query_evolution",
     ],
@@ -77,7 +77,7 @@ pkg_tar(
         "//:busclient",
         "//:busnode",
         "//:database_adapter",
-        "//:tests_db_loader",
+        "//:db_loader",
         "//:word_query",
         "//:word_query_evolution",
     ],
